+++ src/Product/Models/ProductState.php
@@ -17,6 +17,19 @@ namespace Vanilo\Product\Models;
 use Konekt\Enum\Enum;
 use Vanilo\Product\Contracts\ProductState as ProductStateContract;
 
+/**
+ * @property bool $is_draft
+ * @property bool $is_inactive
+ * @property bool $is_active
+ * @property bool $is_unavailable
+ * @property bool $is_retired
+ *
+ * @method static ProductState DRAFT()
+ * @method static ProductState INACTIVE()
+ * @method static ProductState ACTIVE()
+ * @method static ProductState UNAVAILABLE()
+ * @method static ProductState RETIRED()
+ */
 class ProductState extends Enum implements ProductStateContract
 {
     public const __DEFAULT = self::DRAFT;
@@ -27,19 +40,13 @@ class ProductState extends Enum implements ProductStateContract
     public const UNAVAILABLE = 'unavailable';
     public const RETIRED = 'retired';
 
-    protected static $activeStates = [self::ACTIVE];
+    protected static array $activeStates = [self::ACTIVE];
 
-    /**
-     * @inheritdoc
-     */
     public function isActive(): bool
     {
         return in_array($this->value, static::$activeStates);
     }
 
-    /**
-     * @inheritdoc
-     */
     public static function getActiveStates(): array
     {
         return static::$activeStates;