+++ src/Product/resources/database/migrations/2022_02_28_160230_add_original_price_field_to_products_table.php
@@ -1,0 +1,23 @@
+<?php
+
+declare(strict_types=1);
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+
+class AddOriginalPriceFieldToProductsTable extends Migration
+{
+    public function up()
+    {
+        Schema::table('products', function (Blueprint $table) {
+            $table->decimal('original_price', 15, 4)->nullable()->after('price');
+        });
+    }
+
+    public function down()
+    {
+        Schema::table('products', function (Blueprint $table) {
+            $table->dropColumn('original_price');
+        });
+    }
+}