+++ Changelog.md
@@ -15,6 +15,7 @@
 - BC: Renamed `vanilo.framework.*` config values to `vanilo.foundation.*`
 - BC: Added the static `findBySlug(string $slug): ?Property;` method to the Property interface
 - Added Shipment Module
+- Added `original_price` field to products
 
 ## 2.x Series